+++ tpl/academy-payback.js
@@ -1,0 +1,3 @@
+import Parent from './_dummy.js';
+
+export default class PrepareTpl extends Parent {}